Refactor create-agent-dialog.tsx: extract hook and type select
Split create-agent-dialog.tsx (760 → 357 lines) by extracting:
- use-create-agent-form.ts (377 lines): all form state, effects,
  callbacks, and the submit handler
- agent-type-select.tsx (119 lines): self-contained agent type
  dropdown combobox with its own open/close state

The dialog component is now a thin render shell that delegates
to the hook for logic and the AgentTypeSelect for the dropdown.
