+++ apps/web/src/components/app/use-create-agent-form.ts
@@ -1,0 +1,377 @@
+import {
+  type ClipboardEvent,
+  type DragEvent,
+  type FormEvent,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
+import { toast } from "sonner";
+
+import {
+  createClipboardSuggestionFromText,
+  getClipboardFilesFromEvent,
+  startupFileKey,
+} from "@/components/app/create-agent-dialog-clipboard";
+import {
+  CONTEXT_PROMPT_ID,
+  LAST_USED_CWD_KEY,
+  LAST_USED_TYPE_KEY,
+  readLastUsedAgentType,
+  readLastUsedCwd,
+  useCwdHistory,
+  useCreateAgentPrefs,
+} from "@/components/app/create-agent-dialog-utils";
+import { type Agent } from "@/components/app/types";
+import { useSystemDefaults } from "@/hooks/use-system-defaults";
+import { type AgentType } from "@/lib/agent-types";
+import { api } from "@/lib/api";
+
+type UseCreateAgentFormOptions = {
+  enabledAgentTypes: AgentType[];
+  initialAgentType: AgentType | null;
+  resolveDefaultCwd: () => string;
+  onCreated: (agent: Agent, agentType: AgentType) => Promise<void>;
+};
+
+export function useCreateAgentForm({
+  enabledAgentTypes,
+  initialAgentType,
+  resolveDefaultCwd,
+  onCreated,
+}: UseCreateAgentFormOptions) {
+  const [step, setStep] = useState<"config" | "context">("config");
+  const promptTextareaRef = useRef<HTMLTextAreaElement>(null);
+  const [createName, setCreateName] = useState("");
+  const [createType, setCreateType] = useState<AgentType>(() => {
+    const preferred = initialAgentType ?? readLastUsedAgentType();
+    return preferred && enabledAgentTypes.includes(preferred)
+      ? preferred
+      : (enabledAgentTypes[0] ?? "codex");
+  });
+  const [createCwd, setCreateCwd] = useState(() => {
+    const resolved = resolveDefaultCwd().trim();
+    return resolved || readLastUsedCwd();
+  });
+  const [createCwdInitialized, setCreateCwdInitialized] = useState(
+    () => createCwd.trim().length > 0
+  );
+  const [createUseWorktree, setCreateUseWorktree] = useState(true);
+  const [createWorktreeBranch, setCreateWorktreeBranch] = useState("");
+  const [cwdIsGitRepo, setCwdIsGitRepo] = useState<boolean | null>(null);
+  const cwdPathInfoRef = useRef<{ isGitRepo: boolean } | null>(null);
+  const [initialPrompt, setInitialPrompt] = useState("");
+  const [startupFiles, setStartupFiles] = useState<File[]>([]);
+  const [startupLinks, setStartupLinks] = useState<string[]>([]);
+  const [draggingFiles, setDraggingFiles] = useState(false);
+  const [contextDraftInvalid, setContextDraftInvalid] = useState(false);
+  const [creating, setCreating] = useState(false);
+  const {
+    history: cwdHistory,
+    removableHistory: removableCwdHistory,
+    historyMetadata: cwdHistoryMetadata,
+    add: addCwdHistory,
+    remove: removeCwdHistory,
+  } = useCwdHistory();
+  const {
+    fullAccess: createFullAccess,
+    setFullAccess: setCreateFullAccess,
+    autoReview: createAutoReview,
+    setAutoReview: setCreateAutoReview,
+    baseBranch: createBaseBranch,
+    setBaseBranch: setCreateBaseBranch,
+    createNewBranch,
+    setCreateNewBranch,
+  } = useCreateAgentPrefs(createCwd);
+
+  useEffect(() => {
+    setCreateWorktreeBranch("");
+  }, [createCwd]);
+
+  const { data: systemDefaults, isError: systemDefaultsError } =
+    useSystemDefaults();
+  useEffect(() => {
+    if (createCwdInitialized) return;
+    if (systemDefaults) {
+      setCreateCwd(systemDefaults.homeDir);
+      setCreateCwdInitialized(true);
+    } else if (systemDefaultsError) {
+      setCreateCwdInitialized(true);
+    }
+  }, [createCwdInitialized, systemDefaults, systemDefaultsError]);
+
+  useEffect(() => {
+    if (enabledAgentTypes.includes(createType)) return;
+    setCreateType(enabledAgentTypes[0] ?? "codex");
+  }, [createType, enabledAgentTypes]);
+
+  useEffect(() => {
+    if (step === "context") {
+      requestAnimationFrame(() => promptTextareaRef.current?.focus());
+    }
+  }, [step]);
+
+  useEffect(() => {
+    if (step !== "context") {
+      setDraggingFiles(false);
+    }
+  }, [step]);
+
+  const handlePathInfoChange = useCallback(
+    (info: { isGitRepo: boolean } | null) => {
+      cwdPathInfoRef.current = info;
+      setCwdIsGitRepo(info ? info.isGitRepo : null);
+    },
+    []
+  );
+
+  const startupFilePreviewsRef = useRef<Map<string, string>>(new Map());
+
+  const appendStartupFiles = useCallback((files: File[]) => {
+    if (files.length === 0) return;
+    setStartupFiles((current) => {
+      const next = [...current];
+      const seen = new Set(current.map(startupFileKey));
+      for (const file of files) {
+        const key = startupFileKey(file);
+        if (seen.has(key)) continue;
+        seen.add(key);
+        next.push(file);
+        if (
+          file.type.startsWith("image/") &&
+          !startupFilePreviewsRef.current.has(key)
+        ) {
+          startupFilePreviewsRef.current.set(key, URL.createObjectURL(file));
+        }
+      }
+      return next;
+    });
+  }, []);
+
+  const handleAddLink = useCallback((url: string) => {
+    setStartupLinks((current) =>
+      current.includes(url) ? current : [...current, url]
+    );
+  }, []);
+
+  const handleStartupPaste = useCallback(
+    (event: ClipboardEvent<HTMLElement>) => {
+      const target = event.target;
+      const targetIsPrompt =
+        target instanceof HTMLElement && target.id === CONTEXT_PROMPT_ID;
+      if (!targetIsPrompt) return;
+
+      const pastedFiles = getClipboardFilesFromEvent(event);
+      if (pastedFiles.length > 0) {
+        event.preventDefault();
+        appendStartupFiles(pastedFiles);
+        return;
+      }
+
+      const textSuggestion = createClipboardSuggestionFromText(
+        event.clipboardData.getData("text/plain")
+      );
+      if (textSuggestion?.kind === "url") {
+        event.preventDefault();
+        handleAddLink(textSuggestion.url);
+      }
+    },
+    [appendStartupFiles, handleAddLink]
+  );
+
+  const handleStartupDrop = useCallback(
+    (event: DragEvent<HTMLElement>) => {
+      const droppedFiles = Array.from(event.dataTransfer.files ?? []);
+      if (droppedFiles.length === 0) return;
+      event.preventDefault();
+      setDraggingFiles(false);
+      appendStartupFiles(droppedFiles);
+    },
+    [appendStartupFiles]
+  );
+
+  const handleRemoveStartupFile = useCallback((fileToRemove: File) => {
+    const key = startupFileKey(fileToRemove);
+    const url = startupFilePreviewsRef.current.get(key);
+    if (url) {
+      URL.revokeObjectURL(url);
+      startupFilePreviewsRef.current.delete(key);
+    }
+    setStartupFiles((current) =>
+      current.filter((file) => startupFileKey(file) !== key)
+    );
+  }, []);
+
+  useEffect(() => {
+    const previews = startupFilePreviewsRef.current;
+    return () => {
+      for (const url of previews.values()) {
+        URL.revokeObjectURL(url);
+      }
+      previews.clear();
+    };
+  }, []);
+
+  const handleRemoveStartupLink = useCallback((linkToRemove: string) => {
+    setStartupLinks((current) =>
+      current.filter((link) => link !== linkToRemove)
+    );
+  }, []);
+
+  const handleClipboardText = useCallback((text: string) => {
+    setInitialPrompt((current) =>
+      current.trim().length === 0 ? text : `${current.trimEnd()}\n\n${text}`
+    );
+    requestAnimationFrame(() => promptTextareaRef.current?.focus());
+  }, []);
+
+  const enterContextStep = useCallback(() => {
+    setStep("context");
+  }, []);
+
+  const handleSubmit = useCallback(
+    async (event: FormEvent<HTMLFormElement>) => {
+      event.preventDefault();
+      const cwd = createCwd.trim();
+      if (!cwd) return;
+
+      setCreating(true);
+      try {
+        const latestPathInfo = cwdPathInfoRef.current;
+        const submitUseWorktree =
+          latestPathInfo?.isGitRepo === false ? false : createUseWorktree;
+        const contextInitialPrompt =
+          step === "context" ? initialPrompt.trim() || undefined : undefined;
+        const payloadBase = {
+          name: createName.trim(),
+          cwd,
+          type: createType,
+          fullAccess: createFullAccess,
+          autoReview: createAutoReview,
+          useWorktree: submitUseWorktree,
+          createNewBranch: submitUseWorktree ? createNewBranch : undefined,
+          worktreeBranch:
+            submitUseWorktree && createNewBranch
+              ? createWorktreeBranch.trim() || undefined
+              : undefined,
+          baseBranch:
+            submitUseWorktree && createBaseBranch !== "main"
+              ? createBaseBranch
+              : undefined,
+          initialPrompt: contextInitialPrompt,
+        };
+        const resolvedStartupLinks = startupLinks;
+        const useStartupContext =
+          step === "context" &&
+          (startupFiles.length > 0 || resolvedStartupLinks.length > 0);
+        const payload = useStartupContext
+          ? await (async () => {
+              const formData = new FormData();
+              for (const [key, value] of Object.entries(payloadBase)) {
+                if (value === undefined || value === "") continue;
+                formData.append(key, String(value));
+              }
+              formData.append(
+                "startupLinks",
+                JSON.stringify(resolvedStartupLinks)
+              );
+              for (const file of startupFiles) {
+                formData.append("startupFiles", file);
+              }
+              return api<{ agent: Agent }>("/api/v1/agents", {
+                method: "POST",
+                body: formData,
+              });
+            })()
+          : await api<{ agent: Agent }>("/api/v1/agents", {
+              method: "POST",
+              body: JSON.stringify(payloadBase),
+            });
+
+        if (typeof window !== "undefined") {
+          window.localStorage.setItem(LAST_USED_CWD_KEY, cwd);
+          window.localStorage.setItem(LAST_USED_TYPE_KEY, createType);
+        }
+        addCwdHistory(cwd);
+        await onCreated(payload.agent, createType);
+      } catch (err) {
+        const message =
+          err instanceof Error ? err.message : "Failed to create agent.";
+        toast.error(message);
+      } finally {
+        setCreating(false);
+      }
+    },
+    [
+      createAutoReview,
+      createBaseBranch,
+      createCwd,
+      createFullAccess,
+      createName,
+      createNewBranch,
+      createType,
+      createUseWorktree,
+      createWorktreeBranch,
+      initialPrompt,
+      addCwdHistory,
+      onCreated,
+      startupFiles,
+      startupLinks,
+      step,
+    ]
+  );
+
+  const worktreeAvailable = cwdIsGitRepo === true;
+  const worktreeChecked = worktreeAvailable && createUseWorktree;
+
+  return {
+    step,
+    setStep,
+    promptTextareaRef,
+    createName,
+    setCreateName,
+    createType,
+    setCreateType,
+    createCwd,
+    setCreateCwd,
+    createUseWorktree,
+    setCreateUseWorktree,
+    createWorktreeBranch,
+    setCreateWorktreeBranch,
+    initialPrompt,
+    setInitialPrompt,
+    startupFiles,
+    startupLinks,
+    draggingFiles,
+    setDraggingFiles,
+    contextDraftInvalid,
+    setContextDraftInvalid,
+    creating,
+    createFullAccess,
+    setCreateFullAccess,
+    createAutoReview,
+    setCreateAutoReview,
+    createBaseBranch,
+    setCreateBaseBranch,
+    createNewBranch,
+    setCreateNewBranch,
+    cwdHistory,
+    removableCwdHistory,
+    cwdHistoryMetadata,
+    removeCwdHistory,
+    worktreeAvailable,
+    worktreeChecked,
+    startupFilePreviewsRef,
+    handlePathInfoChange,
+    appendStartupFiles,
+    handleAddLink,
+    handleStartupPaste,
+    handleStartupDrop,
+    handleRemoveStartupFile,
+    handleRemoveStartupLink,
+    handleClipboardText,
+    enterContextStep,
+    handleSubmit,
+  };
+}
